Add iocccsize_test.1
It does not include the best action if a problem occurs because I am not
exactly sure what the best action would be aside from probably reporting
it and not relying on it for submitting an entry to the contest. This
can happen on different platforms or even the same platform. For example
it was broken on my macOS system at one point even though it worked fine
under Landon's macOS system. This can be discussed and the man page
updated at that point.