Add ioccc_test.1 + typo fix ioccc_test.sh
Once again I have not said in the man page what action to take when one
or more test fails because I am not sure what that action should be (and
it probably depends on which test failed). This can be discussed on
GitHub.

Another thing to note is that I did not complete SEE ALSO section. This
is because not all of the used tools have man pages so this has to be
fixed later on once that's the case. Unfortunately I am getting to the
point where I am done for the day so it will have to be finished at a
later time and probably date.