Move several man pages to section 8
The following man pages were made section 8 from section 1:

	renamed:    have_timegm.1 -> have_timegm.8
	renamed:    ioccc_test.1 -> ioccc_test.8
	renamed:    iocccsize_test.1 -> iocccsize_test.8
	renamed:    limit_ioccc.1 -> limit_ioccc.8
	renamed:    run_usage.1 -> run_usage.8
	renamed:    utf8_test.1 -> utf8_test.8
	renamed:    verge.1 -> verge.8

The Makefile has been updated for this including making the proper
directories first - a forgotten step almost certainly my doing. An
unfortunate thing that probably won't be a problem for anyone but me as
long as we keep it fixed before public consumption is that these man
pages were already installed under man section 1 so trying to do man on
it will result first in section 1. Perhaps we need an uninstall which
could try removing all man pages (${MANPAGES} in the Makefile) from all
man directories, obviously with -f and redirecting stderr to /dev/null.
If so this can be done later.

I also fixed a typo (well not really a typo but made it more consistent
and 'proper') in have_timegm.c (1st -> first).