Change -p jparse in jparse_test to -j jparse
Since -j was not used it seems like it's more appropriate to use -j for
the json parser. I believe that part of why -p was used (if not entirely
why) is that earlier on we had jinfochk and jauthchk which used -j and
-J and so we used -p for other options though at the time jparse was
finished the tools were already renamed (not sure if it was changed to
just one tool yet though I - and probably Landon as well - had thought
of it being necessary much earlier on).

The man page was updated to account for this.

Tested make test to make sure nothing broke with this change.

It's possible I've missed something but I think everything should be in
order. If not I will fix tomorrow or the next day as I will be unable to
do so a bit later on today and won't be back until tomorrow.

I did also have some fun with oebxergfB.h at the same time as this
commit as well but that doesn't really matter much since only Landon and
I can possibly understand it! :-)