Fix potential invocation error in run_{flex,bison}
The scripts used command -v to determine the path of the tools. But this
is a problem because it doesn't force checking of the path. For example
on my macOS system doing command -v ls reports:

    alias ls='ls -GF'

but with type -P ls:

    /bin/ls

It's improbable that most would have an alias for flex or bison but
better safe than sorry. At this time no other scripts use command but
instead use type (if necessary). However the Makefile does use it but
I'm not sure if this is good or bad. Probably it should also be fixed
but for now I'm leaving it off pending discussion.