Improve (and fix in part) sum_and_check()
The second and third checks on count increment seemed to be in error and
this was fixed but with the caveat that I could be missing something.
There still exists a problem with the summing part of the code but I'm
not as certain what needs to be tested and what does not.

I added more information to all the calls to warn() to help identify any
problems. In my changes (that cannot be committed until sum_and_count()
is fixed) of txzchk.c which uses sum_and_count() one might see the
following output, to give an idea of what it looks like:

    Performing checks on tarball ...
    test-0
    txzchk: ./test_txzchk/good/entry.test-0.19248687209.txt size of 506 bytes OK
    drwxr-xr-x  0 501    20          0 Feb 14 11:58 test-0/
    -rw-r--r--  0 501    20        896 Feb 14 11:58 test-0/Makefile
    -rw-r--r--  0 501    20          4 Feb 14 11:58 test-0/extra2
    -rw-r--r--  0 501    20          0 Feb 14 11:58 test-0/prog.c
    -rw-r--r--  0 501    20       1320 Feb 14 11:58 test-0/.author.json
    -rw-r--r--  0 501    20        778 Feb 14 11:58 test-0/remarks.md
    -rw-r--r--  0 501    20       1258 Feb 14 11:58 test-0/.info.json
    -rw-r--r--  0 501    20          4 Feb 14 11:58 test-0/extra1
    Warning: sum_and_count: third check on sum failed: -sum -1262 != prev_sum-(~inv_value) -1254
    Warning: sum_and_count: third check on sum failed: -sum -782 != prev_sum-(~inv_value) -774
    Warning: sum_and_count: third check on sum failed: -sum -1324 != prev_sum-(~inv_value) -1316
    Warning: sum_and_count: third check on sum failed: -sum -4 != prev_sum-(~inv_value) 4
    Warning: sum_and_count: third check on sum failed: -sum -8 != prev_sum-(~inv_value) 0
    Warning: sum_and_count: third check on sum failed: -sum -900 != prev_sum-(~inv_value) -892
    Warning: sum_and_count: third check on sum failed: -sum -4 != prev_sum-(~inv_value) 4
    Warning: txzchk: ./test_txzchk/good/entry.test-0.19248687209.txt: found 7 feathers stuck in the tarball

As can be seen this is really strange and because I'm not sure what the
purpose of the third test:

     if (-sum != prev_sum-(~inv_value)) {

I will not at this time work on it. The check above it is:

     if ((*sum_checkp)-value != -sum) {

but this makes me more unsure what the purpose of the test is.
Nevertheless this commit should hopefully give more information for
future reference and also fix the count checks (at least those above
the sum check that fails). If that count check needs to be refined more
as well it can be done later but I gave the logic in a comment on
GitHub behind it.