New version of txzchk - address much of #334
Happy Hobbit Day! In honour of both Bilbo Baggins and Frodo Baggins and
their birthdays (which also happened to be a Thursday in the Third Age
3001 or T.A. 3001 or in the Shire Reckoning S.R. 1401 - where Bilbo
turned 111 and Frodo turned 33 which was his coming of age and his
inheritance) I have made some fixes and enhancements to txzchk (okay -
it's not done in honour of the Bagginses but it's fun to think about and
they should be honoured), addressing a number of points in #334.

The new version addresses comments
https://github.com/ioccc-src/mkiocccentry/issues/334#issuecomment-1242821987
and
https://github.com/ioccc-src/mkiocccentry/issues/334#issuecomment-1242823960
as well as later discussion. To be more more specific:

- rounded file size is no longer used.
- there is a limit on the number of files allowed in the tarball
  (currently 42 or MAX_FILE_COUNT).
- there is a limit on the size of all files summed (currently 27651*1024
  or MAX_SUM_FILELEN).
- only regular files count against files and file sizes (so directories,
  sockets and anything else not a regular file will not count against
  file count or total size of all files).
- and possibly some other things I have forgotten.

As well I made the following changes:

- More debug information (show the size of required files for example).
- Increased the verbosity level of some debug messages (this is because
  of debug information added to util.c in sum_and_count() - the
  information in txzchk is not useful to have at medium debug level).
- Fixed the case where it would say 'files' even when it was a single
  file.
- A new 'static char const *tok_sep = " \t";' used instead of having " \t"
  strewn about the code (instead 'tok_sep' can be strewn about the code! :-) ).
- Each file in the files list has a size member (this was originally
  going to be used for something but it was not needed and I only see it
  now with git diff so not removing it: it might be removed in the
  future should it prove useless).
- Probably some other changes in txzchk.c as well.

I'm aware of the issue of part of sum_and_count() might better off use off_t
but for now it uses intmax_t.

Fix string_to_intmax2() to check for non-digits after digits as per
comment
https://github.com/ioccc-src/mkiocccentry/issues/334#issuecomment-1242821987
which shows the invalid line:

    -rw-r--r--  0 501    20        4.0 Jun  4 04:52 12345678-1234-4321-abcd-1234567890ab-2/extra2

which previously was valid because strtoimax() will not report an error
there. Whether the other string_to_ functions need to be updated for
this (including string_to_intmax()) I am uncertain but I wanted to get
these changes in.

Temporarily comment out the test in sum_and_count() that has a
calculation error with a 'FIXME' note on it. This is so the changes to
txzchk could be tested successfully (as I noted yesterday I'm not sure
the purpose of that specific check so I can't fix it without risking
losing an important test that is necessary in the judging process).

Added three files to test_txzchk/bad and one to test_txzchk/good (all
pass).

Updated test_JSON tree so that all cases of the previous txzchk version
was changed to the new version.