Improve and bug fix jstr_test.sh
Now allows for specifying path to jstrencode and jstrdecode. The usage
looks like:

    usage: ./jstr_test.sh [-h] [-v level] [-e jstrencode] [-d jstrdecode]

	-h		print help and exit 1
	-v		set verbosity level for this script (def level: 0)
	-e		path to jstrencode tool (def: ./jstrencode)
	-d		path to jstrdecode tool (def: ./jstrdecode)

    exit codes:

	0	    - all is well
	1	    - -h used
	2	    - invalid option
	3	    - option argument missing
	42	    - test #0 failed
	43	    - test #1 failed
	44	    - test #2 failed
	45	    - test #3 failed
	100	    - jstrencode not executable
	101	    - jstrdecode not executable
	102	    - jstr_test.out exists and can't remove it
	103	    - jstr_test2.out exists and can't remove it

The script first tries removing the test files and if they exist after
that then it fails whereas previously it was a fatal error if they
existed (saying to move or remove them).

At least two exit codes were used for more than one purpose - in getopts
parsing and also as a test failure code. This has been rectified.

The exit codes have been changed a bit so that more tests could be added
if necessary (or any other category of exit codes). The command line /
getopts codes currently are 1 - 3; the exit codes for test failures are
42 - 45 and the checks prior to running the tests are 100 - 103.

Passes shellcheck.

The man page still needs to be created and I plan on doing this in the
near future - possibly today but more likely it won't be.