Add bad test file to test_txzchk
The output should look like:

    $ ./txzchk -T -E txt test_txzchk/bad/entry.test-0.1662145368.txt
    Welcome to txzchk version: 0.12 2022-09-22

    Performing sanity checks on your environment ...
    ... environment looks OK

    Performing checks on tarball ...
    test-0
    txzchk: test_txzchk/bad/entry.test-0.1662145368.txt size of 765 bytes OK
    drwxr-xr-x  0 501    20          0 Jul 28 08:26 test-0/
    -rw-r--r--  0 501    20          0 Jul 28 08:25 test-0/file bogus data
    -rw-r--r--  0 501    20       1856 Jul 23 07:44 test-0/Makefile
    -rw-r--r--  0 501    20          4 Jul 23 07:44 test-0/extra2
    -rw-r--r--  0 501    20          0 Jul 23 07:44 test-0/prog.c
    -rw-r--r--  0 501    20       2793 Jul 23 07:44 test-0/.author.json
    -rw-r--r--  0 501    20       3193 Jul 23 07:44 test-0/remarks.md
    -rw-r--r--  0 501    20       1430 Jul 23 07:44 test-0/.info.json
    -rw-r--r--  0 501    20          4 Jul 23 07:44 test-0/extra1
    -rw-r--r--  0 501    20          4 Jul 23 07:44 test-0/./extra1
    -rw-r--r--  0 501    20          4 Jul 23 07:44 curds/whey/extra
    -rw-r--r--  0 501    20          4 Jul 23 07:44 curds/file
    Warning: txzchk: test_txzchk/bad/entry.test-0.1662145368.txt: found incorrect directory in filename curds/file
    Warning: txzchk: test_txzchk/bad/entry.test-0.1662145368.txt: found more than one directory in path curds/whey/extra
    Warning: txzchk: test_txzchk/bad/entry.test-0.1662145368.txt: found incorrect directory in filename curds/whey/extra
    Warning: txzchk: test_txzchk/bad/entry.test-0.1662145368.txt: found more than one directory in path test-0/./extra1
    Warning: txzchk: test_txzchk/bad/entry.test-0.1662145368.txt: bogus field found after filename: bogus
    Warning: txzchk: test_txzchk/bad/entry.test-0.1662145368.txt: bogus field found after filename: data
    Warning: txzchk: test_txzchk/bad/entry.test-0.1662145368.txt: found a total of 2 files with the name extra1
    Warning: txzchk: test_txzchk/bad/entry.test-0.1662145368.txt: found directory change in filename curds/whey/extra
    Warning: txzchk: test_txzchk/bad/entry.test-0.1662145368.txt: found directory change in filename curds/file
    Warning: txzchk: test_txzchk/bad/entry.test-0.1662145368.txt: found a total of 2 files with the name file
    Warning: txzchk: test_txzchk/bad/entry.test-0.1662145368.txt: not all files in correct directory
    Warning: txzchk: test_txzchk/bad/entry.test-0.1662145368.txt: found 11 feathers stuck in the tarball

However I seem to recall that although invalid technically, I had
originally programmed '././' to not count against extra directories (it
would still be disqualified for screwing with the tarball). I thought it
should be the same for 'test-0/./' but perhaps I did not do that after
all.

A word about the bogus fields. This simplifies the code a fair bit by
not needing to worry about trying to find files with spaces. As those
would be rejected anyway we can just say that the tarball has bogus
fields after the filename even if in some cases the filename includes
those bogus fields. This was discussed on the GitHub page for txzchk but
it's worth pointing out again anyway as at first glance one might think
it should report that it's a file with spaces - but can we be sure
that's what is actually happening? Since we can't I decided to just
report it as a bogus field. As for why report it as two bogus fields
it's because the separator for strtok_r() is " \t" as is necessary.

The above two paragraphs is commentary that probably is only relevant
for the IOCCC Judges but I am including it anyway as it's shown in this
commit.