Make make picky run picky on shell scripts
One file has changed which is because of a space at the end of the line.

I did not put a line length limit because there are some scripts that
have long lines with 'echo' and breaking them into more than one line
and getting the same output would mean moving the next line(s) to the
first column which makes indentation inconsistent (there are other
solutions but those cause different kinds of problems). The options
specified in particular are: -w -u -s -t8 -v -e -8.