`make parser' now compiles chkentry
The question of whether or not it should just do `make all' is certainly
something to consider but for now this covers everything that will need
to be recompiled after the parser is regenerated.

There was also a typo fix in jparse.l - kind of. I had removed 'demo'
from 'demo tool' but I meant to add it back as maybe jparse itself is
kind of a demo as it doesn't actually do anything with it - it just
parses it but does nothing with it other than report if it's valid or
not.