Bug fix jstr_test.sh and improve jstr_test.8
There was a reference (or two counting the echo) in jstr_test.sh to
./jstrdecode when it should have been $JSTRDECODE.

The man page is more detailed on what the script does and why it's
important.

This update was actually accidental as I had forgotten I already added
it. This happened as follows: I looked at the GitHub page but after this
man page had been added a new comment apparently was not made so I
looked at the old list. I did not check if the man page was there and
instead copied another script man page to this one and then edited it as
if it did not exist. This turned out for the better though because I
noticed the bug in jstr_test.sh and the man page is better now.