Add jnum_gen.8 and jnum_chk.8 man pages
Running make install will install the two new man pages.

With this commit it seems that the only man pages that need to be
written (as in those that do not exist already) that are JSON specific
are those to do with the JSON semantics tables which are jsemcgen.8 and
jsemtblgen.8. These will be added at a later date.

It also appears that the only other man page to be added (as of this
time) is dyn_array.3 but that will take more time as well.

Many man pages will be changed in due course and some are incomplete
because those tools are incomplete. The two that come to mind are
txzchk_test.8. The txzchk_test.sh script is on my mind to finish but
finish only means make it have less confusing and contradictory output
and add verification that the issues raised are what's expected. I hope
to do this in the near future but I don't know exactly when I will have
the time and energy to do so.

There were also some typo fixes and clarifications in this commit wrt
jnum_gen and jnum_chk. In the jnum_gen help string there was a
formatting issue as well that has been fixed.