Update chkentry.1
As a way to clarify a comment in the log for commit
b8f510ca452c96c0fe8104dfe16f8b44d7e9f34e I have updated chkentry.1. I
had said that I am aware of at least two man pages that are not complete
because the tools in question are incomplete. As I was writing about
txzchk_test.sh (and txzchk_test.8) though I forgot to finish it and say
that the other one is chkentry.1. Thus I have updated chkentry.1 by
saying that because the tool is not finished the man page is not either.