Update txzchk version
As I suggested in commit c5d17b06c5a691b2f66c7eecd90397578c944029 it
felt like it was worthy of a new version but I did not have the energy
or time to do it yesterday. I have done this now but the version is
backdated to yesterday as that's when the commit was. Thus version.h and
many files under test_JSON/ have been updated (using vermod.sh).