Bug fix and typo fix txzchk_test.sh
Fix script so that if run_test has an error the script will exit with
non-zero so that make test will report an issue (as well as the script
itself).

The typo is that outside the function run_test it referred to run_test
(as in the message suggested it was in run_test).