Fix txzchk_test.sh under linux
A problem occurred in string_to_intmax2() in the order of checks for
failures so that under linux the errno was set to non-zero but this was
not triggered under macOS. This fix saves errno first (as an abundance
of caution) and then checks the other conditions. Then after those are
passed the saved errno is checked and if not 0 then it's used (setting
errno to the saved value).

I also updated the test_txzchk/bad error file. This should solve the
problem but if not I'll have to look at it tomorrow.