Improve Makefile rule and hostchk.sh
The Makefile rule now checks the status of the script and reports a
warning if it's not 0.

Added bug-report to .PHONY. There still remains the script bug-report.sh
to write but there is (currently) an overlap of the two scripts and
right now the rule also will run with bug-report so it should be in
.PHONY.

Add -V option to script and print the version in the USAGE string as
well. As noted in commit ebfc386033db38b363cb558e039c7da0ed1a1b57 many
other scripts have to do this as well. This is part of working on
hostchk.sh and bug-report.sh (again not written yet).

Improve the error message when a header file is missing. It previously
looked strange as it showed e.g.:

    FATAL: missing system include file <#include <stdio.h>>"

whereas now it just says it can't compile the code with #include
<stdio.h>.