Fixes to have_timegm.{sh,8}
Add -V option to print version.

Add USAGE string which now prints the version as well (the other scripts
need to do this also).

The script now returns the exit code of the have_timegm.c program so
that the script exit codes can be consistent with the program itself.
This means that other exit codes in the script were changed.

The man page had exit codes related to reset_tstamp.8 for some reason.

Added the -V option to the man page as well.