Add -V option to txzchk_test.sh
The version is printed in the usage string. The number was arbitrarily
selected as a starting point. There have been a number of revisions to
it but whether 3 (i.e. 4) is the correct number or not I do not know. I
backdated it to yesterday which was the most recent change in
functionality.

Updated txzchk_test.8 to account for -V option.