Add pre-commit.sample hook to help with commits
This is completely optional! It is only useful if you'll be making major
changes. There are some caveats. For example if a test fails because of
a platform specific problem but you have other changes to make (but
cannot fix the problem) then this script will pose a problem. In this
case you can temporarily rename or delete the file under the hooks
directory to disable it. For most people it's probably not even useful
but I'm committing it for discussion and/or improvements. This might or
might not be worth even having in the repo but time will tell on that.

Again it is not required but for those making major changes it can be
useful. To install it copy 'pre-commit.sample' to
'.git/hooks/pre-commit'. That's all there is to it.

It looks like:

    #!/bin/sh
    #
    # Sample hook for pre-commit that checks if make test will pass before
    # committing. This is not required for making pull requests but it will prevent
    # having to fix any changes that are merged prior to being committed.
    #
    # To install copy this script as pre-commit to the .git/hooks subdirectory in
    # your clone of your fork.
    #
    # Again this is completely optional but recommended if you'll be making
    # significant changes. In the case that one or more tool changes because of a
    # platform specific problem or a change you did not make then this might pose a
    # problem and that is one of the reasons it is completely optional. Nevertheless
    # it can help you discover a problem with changes you made prior to committing.

    if ! make test; then

	    cat <<\EOF
    Error: Attempt to commit when make test fails.
    EOF
	    exit 1
    fi