Add bug_report.sh
For now it's just a stub that exits 0. It is currently being discussed
and later on (probably at another date) I will start working on it. I
added it here because the Makefile refers to it.

The Makefile has been updated so that when the script has some code it
will be checked via shellcheck. Also the name was changed to
bug_report.sh to match the style of the rest of the files (some of which
used to have - but were changed to underscore).