Update bug_report.sh - initial code
The script is still quite basic but it now creates a log file, reporting
the name if -v > 1 (and always at the end of the script). It then runs,
noting each step in the log file as well as stdout, make clobber, make
all and make test.

If any step fails it changes the exit code and notes this in the log
file as well. At the end if the exit code is not 0 it suggests the user
report an issue and attach the log file. Otherwise it tells the user
that no test fails and lets them know that they can safely delete the
log file if they decide to.

The log file name is calculated like:

    LOG_FILE="bug-report.$(/bin/date +%Y%m%d.%H%M%S).txt"

and the reason it doesn't start with bug_report is to distinguish it
from the script itself.

Added to .gitignore bug-report*.txt.

A lot more has to be added but this is a good start. It's possible I
will add more today but that's not certain. For now I am going to look
at the comments on the GitHub issue page and after replying I'll see how
I feel.