Add call to hostchk.sh in bug_report.sh
There is also a note of a bug in the script with a FIXME note.

And no: amusing as it would have been I cannot say this was intentional
:-) I did think of it earlier on but I forgot to address it and this
will not be fixed until another date. Still it's probably good that it's
there because we can all use some amusement and laughs. Anyway the
problem is that the exit codes will not be tested right which makes
checking for them entirely useless.