Delete README.md
We do not wish to include the  pre-commit.sample file.  While we do understand that the use of this file is optional, we do not intend to use it.  In particular we do not intend to move this file into place, but rather perform a `make release` action and proceed if all is well instead.