Add initial code of bug_report.sh
The script is still quite basic but it now creates a log file, reporting
the name if -v > 1 (and always at the end of the script). It then runs,
noting each step in the log file as well as stdout, make clobber, make
all and make test. It also runs hostchk.sh.

If any step fails it changes the exit code and notes this in the log
file as well (but see below). At the end if the exit code is not 0 it
suggests the user report an issue and attach the log file. Otherwise it
tells the user that no test fails and lets them know that they can
safely delete the log file if they decide to.

The log file name is calculated like:

    LOG_FILE="bug-report.$(/bin/date +%Y%m%d.%H%M%S).txt"

and the reason it doesn't start with bug_report is to distinguish it
from the script itself.

Added to .gitignore bug-report*.txt.

There is also a note of a bug in the script with a FIXME note.

And no: amusing as it would have been I cannot say this was intentional
:-) I did think of it earlier on but I forgot to address it and this
will not be fixed until another date. Still it's probably good that it's
there because we can all use some amusement and laughs. Anyway the
problem is that the exit codes will not be tested right which makes
checking for them entirely useless.

The Makefile has been updated so that make shellcheck will check the
script. Also changed the name to bug_report.sh to match the style of the
rest of the files (some of which used to have - but were changed to
underscore).