Bug fix bug_report.sh :-) + add bug_report.1
The script now properly detects if a test fails. I've left a comment
about this bug in the script because this world needs more reasons to
laugh and it is funny and ironic that a script that helps one report
bugs actually had a bug that would make it seem like there was no bug.

I updated the version though there's no other real change except that it
gives more information on reporting bugs/issues and even does so if the
script does not detect any problems (in case for example one thinks that
there is a problem but the script did not detect it).