Add hostchk.1 & hostchk.sh + bug_report.sh updates
The update in hostchk.sh (of significance) is a note about how if this
script fails one will be (via the Makefile) be warned about it and that
warning it will recommend one to run bug_report.sh which itself will run
hostchk.sh which will suggest that you run bug_report.sh :-) A similar
note is also in commit 79cef044d6f50601a25eb30a9c4e6fabfc6af4bf.

There is currently no plan to remove this because if the script does
find a problem it is most likely a problem that needs to be addressed if
one is to successfully use the toolkit. So even if it is redundant or
feels like it's a circular dependency (but it's not entirely :-) ) it's
for now going to stay that way to increase the chance that one does not
ignore the warnings.

A note about this has been added to the new hostchk.1 man page as well
as the bug_report.1 man page.

Sorry in advance if (to use an American English phrase which I will
happily use for a pun for a pun is a pun and a pun not made is a wasted
opportunity :-) ) this throws you for a loop. :-) But then again, now
you're in the loop so it should not be a problem. :-)