Fix make checknr in hostchk.1
It appears that if a \fB..\fP pair is on more than one line (due to
being at end of line) checknr has a problem parsing it. Perhaps this is
another reason to use the .B (or .BR) macro in man pages but for now I'm
not doing that as that would take more time. If I do do it it'll be
later down the road and probably all man pages except for the SYNOPSIS
(though there is room for improvement in that too but currently the way
that section is formatted the run_usage.sh works good whereas if I were
to change it it would not).