Fixes and improvements in fnamchk.1
The options are improved in format in that the ones that take an
argument have the actual argument specified (underlined) and the macro
used is better (no longer .PP but .TP).

I also updated the filenames to have the most recent MIN_TIMESTAMP and
added a note that if that value changes and the filenames are not
updated in the man page the example will fail. Of course if an example
is added for incorrect timestamp it should be before this value
(1662145368) as otherwise when (as I suggested in the man page itself)
one does a search and replace it would make the example pass the checks
when it should fail.