Improve fnamchk.1
Add an example where the timestamp is invalid.

Rearrange the order of the examples so that the note in the man page
that refers to the minimum timestamp can be simplified and can assume
that every example below has a valid timestamp. The examples that
will fail because of something else (i.e. invalid format and also with a
timestamp that is no longer valid) are above the note.