Change signal_spec to EXIT from list of signals
In three scripts that use trap it trapped specific signals but I
suggested that we trap 'EXIT' instead. This was agreed to be a better
idea so I have changed the trap signal_spec to be EXIT in jsemcgen.sh,
jstr_test.sh and old.chkcode.sh (txzchk_test.sh already had EXIT).