Fixes and clean up in fnamchk.1
Reword the EXAMPLES section so that it no longer needs to be updated
when the minimum timestamp is updated.

I also changed the formatting a bit so that it should be cleaner. This
same change, which is moving some macros about, needs to be done with
the other man pages which have an EXAMPLES section but this might not be
the highest priority right now - I only did it because I was fixing the
section anyway. Nevertheless it might be a good idea to do it sooner
than later.