Improve bug_report.sh reporting
The script now shows a summary of issues detected (just like
ioccc_test.sh) does.

Additionally it writes the final exit code (if non-zero) to the log file
and it always writes the final notice (whether any issues were found or
not) to the log file including the name of the log file. This is so that
when the user uploads the log file we can verify that they actually
uploaded the correct file for the issues being reported.

And perhaps I did not do the chmod -x on the hostchk.sh as I recall only
doing it on one script and bug_report.sh was also -x. Perhaps I did but
anyway it's now been fixed.