New version of bug_report.sh - many updates
This version has many additions and improvements. Since there are
probably fewer improvements I'll document those first.

Instead of trying to keep track of the exit code each time a test fails
we start the exit code at 10. Then if a test fails it will be
incremented without any concern of what the previous value was. This is
because the only reason the exit code matters is to indicate that a
problem was detected. The error message is what tells us what really
went wrong and this way we can rearrange tests, add new tests in the
middle etc. (as some tests might be thought of later on but which should
be done before other tests) without having to then update all the other
references. At the end if the exit code is not 10 (which was arbitrarily
selected) we know that at least one issue was detected. If it's 10 at
the end we set it back to 0 so that the script will not fail when all is
actually okay.  This greatly simplifies adding new tests.

Add function is_exec which tests if a file exists, is a regular file and
is executable. If any of those conditions fail it reports it as an error
and returns 1. This function is used later in the script that acts on
the new TOOLS variable (at the top in case we have to update the list of
tools).

Check that all our tools are available and executable (or at least those
that were referred to in the GitHub comment - it's possible some are
missing from the list but it's easy to update by adding to the TOOLS
variable). If the tool is not executable it is an error that is
reported. Otherwise run -h on the tool. We might only need the version
eventually but this is not decided and not all tools (scripts in
particular) have versions yet.

The script now runs the following commands:

    uname -a
    which cc
    cc -v
    which make
I forgot to check the make version which due to the improvements with
exit codes I can easily add later.

    which tar

We don't need to check that tar has the correct options implemented
because hostchk.sh will do that for us.

These are only what was added in this commit. Much more needs to be done
and it could certainly be made even more modular but this is great
progress. The next step is probably to add simple C files that should be
test compiled as well as running some of the scripts like run_bison.sh
and run_flex.sh. Some of this might come later today or it might be
another day.